feat: Implement ForgeInfra structure and associated services for infrastructure management
