+++ crates/forge_infra/src/lib.rs
@@ -6,7 +6,7 @@ mod fs_read;
 mod fs_remove;
 mod fs_snap;
 mod fs_write;
-mod infra;
+mod forge_infra;
 mod qdrant;
 
-pub use infra::*;
+pub use forge_infra::*;
